type size check

+++ include/ns_type.h
@@ -67,7 +67,7 @@ typedef long i64;
 typedef unsigned char u8;
 typedef unsigned short u16;
 typedef unsigned int u32;
-typedef unsigned long long u64;
+typedef unsigned long u64;
 
 typedef float f32;
 typedef double f64;
